Code refactoring.
Split useNormalizedQuestions, add normalizedQuestionsReducer to the separate file.
Remove array versions of the regular types.
Change generateRandomNumber to use NaN instead of infinity. It's just more convenient.
Change routes to base routes instead of relative. Alternatively we can use baseUrl in ".env" file.
