+++ src/hooks/useNormalizedQuestions.tsx
@@ -1,137 +1,16 @@
 import { useReducer, useEffect } from 'react';
 
 // utils
-import {
-  generateRandomNumber,
-  fetchData,
-  normalizeQuestions,
-  getWordClassName,
-} from '../utils/utils';
+import { generateRandomNumber, fetchData } from '../utils/utils';
 
-// types
+// reducer
 import {
-  NormalizedQuestionArray,
-  NormalizedQuestionWordArray,
-  QuestionArray,
-} from '../utils/types';
-
-// constants
-import { DEFAULT_WORD_CLASS_NAME } from '../utils/constsants';
+  initialState,
+  normalizedQuestionReducer,
+} from './normalizedQuestionsReducer';
 
 interface Props {
   dataUrl: string;
-}
-
-// In future can be moved to the types.ts, however for now we are using it only in this component
-interface NormalizedData {
-  normalizedQuestion: string;
-  normalizedWords: NormalizedQuestionWordArray;
-}
-
-interface ReducerState extends NormalizedData {
-  questions: NormalizedQuestionArray;
-}
-
-type NormalizedQuestionDispatchAction =
-  | {
-      type: 'setQuestions';
-      data: QuestionArray;
-    }
-  | { type: 'setNormalizedQuestionAndWords'; data: NormalizedData }
-  | {
-      type: 'updateClassNameForSelectedWords';
-      isPlaying: boolean;
-    }
-  | {
-      type: 'toggleWordSelection';
-      wordId: string;
-      isPlaying: boolean;
-    };
-
-const initialState: ReducerState = {
-  questions: [],
-  normalizedQuestion: '',
-  normalizedWords: [],
-};
-
-function updateClassNameForSelectedWords(
-  allWords: NormalizedQuestionWordArray,
-  isPlaying: boolean
-) {
-  const normalizedWords = allWords.map((word) => {
-    if (word.selected) {
-      return {
-        ...word,
-        className: getWordClassName(DEFAULT_WORD_CLASS_NAME, word, isPlaying),
-      };
-    }
-
-    return word;
-  });
-
-  return normalizedWords;
-}
-
-function toggleWordSelection(
-  allWords: NormalizedQuestionWordArray,
-  action: { wordId: string; isPlaying: boolean }
-) {
-  const { wordId, isPlaying } = action;
-  const normalizedWords = allWords.map((word) => {
-    if (word.id === wordId) {
-      const newWord = { ...word, selected: !word.selected };
-      const newClassName = getWordClassName(
-        DEFAULT_WORD_CLASS_NAME,
-        newWord,
-        isPlaying
-      );
-
-      return {
-        ...newWord,
-        className: newClassName,
-      };
-    }
-
-    return word;
-  });
-
-  return normalizedWords;
-}
-
-function normalizedQuestionReducer(
-  state: ReducerState,
-  action: NormalizedQuestionDispatchAction
-) {
-  switch (action.type) {
-    case 'setQuestions':
-      return {
-        ...state,
-        questions: normalizeQuestions(action.data, DEFAULT_WORD_CLASS_NAME),
-      };
-
-    case 'setNormalizedQuestionAndWords':
-      return {
-        ...state,
-        ...action.data,
-      };
-
-    case 'updateClassNameForSelectedWords':
-      return {
-        ...state,
-        normalizedWords: updateClassNameForSelectedWords(
-          state.normalizedWords,
-          action.isPlaying
-        ),
-      };
-
-    case 'toggleWordSelection':
-      return {
-        ...state,
-        normalizedWords: toggleWordSelection(state.normalizedWords, action),
-      };
-    default:
-      throw new Error();
-  }
 }
 
 const useNormalizedQuestions = ({ dataUrl }: Props) => {
